refactor: UI fixture 파일 경로 변경
