+++ app/src/main/java/woowacourse/payments/ui/cards/Card.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import woowacourse.payments.ui.cards.core.mapper.asColor
 import woowacourse.payments.ui.model.PaymentCardUiModel
-import woowacourse.payments.ui.preview.paymentCardUiModelSample
+import woowacourse.payments.ui.debug.fixture.paymentCardUiModelSample
 
 @Composable
 fun PaymentCard(
