+++ app/src/main/java/woowacourse/payments/ui/cards/CardsTopBar.kt
@@ -16,8 +16,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import woowacourse.payments.R
 import woowacourse.payments.ui.cards.model.CardsState
-import woowacourse.payments.ui.preview.paymentCardUiModelSample
-import woowacourse.payments.ui.preview.paymentCardUiModelSamples
+import woowacourse.payments.ui.debug.fixture.paymentCardUiModelSample
+import woowacourse.payments.ui.debug.fixture.paymentCardUiModelSamples
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
